rename gpQuoteStatus to gpPriceStrategy
